完善数据保存和恢复方法；修改输入模型的特征

+++ ml_code/main_train.py
@@ -1,7 +1,8 @@
 #!/user/bin/env python
 # -*- coding:utf-8 -*-
 
-from ml_code.data_utils import load_data_from_csv, seg_words
+from ml_code.data_utils import load_data_from_csv, seg_words, get_tfidf_and_save, load_tfidf_dict,\
+    get_vector_tfidf, padding_data
 from ml_code.model import TextClassifier
 from sklearn.feature_extraction.text import TfidfVectorizer
 from ml_code import config
@@ -14,6 +15,8 @@ import argparse
 train_data_path = "../data/sentiment_analysis_trainingset.csv"
 validate_data_path = "../data/sentiment_analysis_validationset.csv"
 models_dir = "models"
+tfidf_path = "data/tfidf.txt"
+max_len = 500
 
 logging.basicConfig(level=logging.INFO, format='%(asctime)s [%(levelname)s] <%(processName)s> (%(threadName)s) %(message)s')
 logger = logging.getLogger(__name__)
@@ -25,9 +28,12 @@ class Main:
         self.train_data_df = None   # 训练集
         self.validate_data_df = None    # 验证集
         self.string_train = None  # 训练集的评论（string）
+        self.string_valid = None    # 验证集的评论（string）
         self.columns = None  # 列索引名称
         self.vectorizer_tfidf = None    # 特征提取器
         self.classifier_dict = None  # 存储所有训练得到的分类器（每种情绪分别训练一个分类器）
+        self.train_data = None  # 用于训练的评论序列
+        self.valid_data = None  # 用于验证的评论序列
 
     def get_parser(self):
         parser = argparse.ArgumentParser()
@@ -41,30 +47,54 @@ class Main:
         logger.info("start load data")
         self.train_data_df = load_data_from_csv(train_data_path)
         self.validate_data_df = load_data_from_csv(validate_data_path)
-        content_train = self.train_data_df.iloc[:, 1]
+        content_train = self.train_data_df.iloc[:2000, 1]    # 训练集评论语句
         logger.info("start seg train data")
-        self.string_train = seg_words(content_train)
-        joblib.dump(self.string_train, "pkl/string_train.pkl")
+        if not os.path.exists("pkl/string_train.pkl"):
+            self.string_train = seg_words(content_train)
+            joblib.dump(self.string_train, "pkl/string_train.pkl")
+        else:
+            self.string_train = joblib.load("pkl/string_train.pkl")
         logger.info("complete seg train data")
+        content_validate = self.validate_data_df.iloc[:, 1]  # 验证集评论语句
+        logger.info("start seg validate data")
+        if not os.path.exists("pkl/string_valid.pkl"):
+            self.string_valid = seg_words(content_validate)
+            joblib.dump(self.string_valid, "pkl/string_valid.pkl")
+        else:
+            self.string_valid = joblib.load("pkl/string_valid.pkl")
+        logger.info("complete seg validate data")
         self.columns = self.train_data_df.columns.values.tolist()   # 获取列索引的名称
 
     def get_feature(self):
-        logger.info("start train feature extraction")
-        self.vectorizer_tfidf = TfidfVectorizer(analyzer='word', ngram_range=(1, 5), min_df=5, norm='l2')
-        self.vectorizer_tfidf.fit(self.string_train)
-        joblib.dump(self.vectorizer_tfidf, "pkl/vectorizer_tfidf.pkl")
-        logger.info("complete train feature extraction models")
-        # logger.info("vocab shape: %s" % np.shape(self.vectorizer_tfidf.vocabulary_.keys()))
+        # logger.info("start train feature extraction")
+        # self.vectorizer_tfidf = TfidfVectorizer(analyzer='word', ngram_range=(1, 5), min_df=5, norm='l2')
+        # self.vectorizer_tfidf.fit(self.string_train)
+        # logger.info("complete train feature extraction models")
+        if not os.path.exists(tfidf_path):
+            get_tfidf_and_save(self.string_train, tfidf_path)
+        tfidf_dict = load_tfidf_dict(tfidf_path)
+        if not os.path.exists("pkl/train_data.pkl"):
+            train_vector_tfidf = get_vector_tfidf(self.string_train, tfidf_dict)
+            self.train_data = padding_data(train_vector_tfidf, max_len)
+            joblib.dump(self.train_data, "pkl/train_data.pkl")
+        else:
+            self.train_data = joblib.load("pkl/train_data.pkl")
+        if not os.path.exists("pkl/valid_data.pkl"):
+            valid_vector_tfidf = get_vector_tfidf(self.string_valid, tfidf_dict)
+            self.valid_data = padding_data(valid_vector_tfidf, max_len)
+            joblib.dump(self.valid_data, "pkl/valid_data.pkl")
+        else:
+            self.valid_data = joblib.load("pkl/valid_data.pkl")
 
     def train(self):
         logger.info("start train model")
         self.classifier_dict = dict()
         f1_score_dict = dict()
         for index, column in enumerate(self.columns[2:3]):
-            label_train = self.train_data_df[column].iloc[:]
+            label_train = self.train_data_df[column].iloc[:2000]
             text_classifier = TextClassifier(vectorizer=self.vectorizer_tfidf)
             logger.info("start train %s model" % column)
-            text_classifier.fit(self.string_train, label_train)
+            text_classifier.fit(self.train_data, label_train)
             logger.info("complete train the %s th model: %s" % (str(index), column))
             f1_score = self.valid(text_classifier, column)
             f1_score_dict[column] = f1_score
@@ -75,13 +105,8 @@ class Main:
         self.valid_all(f1_score_dict)
 
     def valid(self, text_classifier, model_name):
-        content_validate = self.validate_data_df.iloc[:, 1]
-        logger.info("start seg validate data")
-        content_validate = seg_words(content_validate)
-        logger.info("complete seg validate data")
-        logger.info("start validate model")
         label_validate = self.validate_data_df[model_name]
-        f1_score = text_classifier.get_f1_score(content_validate, label_validate)
+        f1_score = text_classifier.get_f1_score(self.valid_data, label_validate)
         return f1_score
 
     def valid_all(self, f1_score_dict):
